Update pom.xml
Added <bundle>true</bundle> which avoids circular dependency errors: "TypeError: Super expression must either be null or a function, not undefined". The fix is to bundle files as mentioned here: 
https://medium.com/visual-development/how-to-fix-nasty-circular-dependency-issues-once-and-for-all-in-javascript-typescript-a04c987cf0de
Of course in this simple Quickstart example where there are so few javascript files and no circular dependency problems, it's unnecessary. But if new users use this quickstart as a working template to convert their own project from Java to Javascript, it's best that we include the best default options to make their experience as painless as possible. When I ran into this circular dependency problem I was close to giving up on JSweet since I had no idea why it didn't work. I happened upon this fix and persevered and am very happy with JSweet. So I would like to contribute this minor edit. Thank you for the great work on JSweet!